fix(OMN-13566): wire no-new-os-environ pre-commit
